+++ src/main/java/org/truffleruby/options/Options.java
@@ -134,6 +134,8 @@ public class Options {
     public final boolean ARGV_GLOBALS;
     /** --gets-loop=false */
     public final boolean GETS_LOOP;
+    /** --print-loop=false */
+    public final boolean PRINT_LOOP;
     /** --ignore-lines-before-ruby-shebang=false */
     public final boolean IGNORE_LINES_BEFORE_RUBY_SHEBANG;
     /** --syntax-check=false */
@@ -324,6 +326,7 @@ public class Options {
         CEXTS_LOG_WARNINGS = options.get(OptionsCatalog.CEXTS_LOG_WARNINGS_KEY);
         ARGV_GLOBALS = options.get(OptionsCatalog.ARGV_GLOBALS_KEY);
         GETS_LOOP = options.get(OptionsCatalog.GETS_LOOP_KEY);
+        PRINT_LOOP = options.get(OptionsCatalog.PRINT_LOOP_KEY);
         IGNORE_LINES_BEFORE_RUBY_SHEBANG = options.get(OptionsCatalog.IGNORE_LINES_BEFORE_RUBY_SHEBANG_KEY);
         SYNTAX_CHECK = options.get(OptionsCatalog.SYNTAX_CHECK_KEY);
         ARGV_GLOBAL_VALUES = options.get(OptionsCatalog.ARGV_GLOBAL_VALUES_KEY);
@@ -506,6 +509,8 @@ public class Options {
                 return ARGV_GLOBALS;
             case "ruby.gets-loop":
                 return GETS_LOOP;
+            case "ruby.print-loop":
+                return PRINT_LOOP;
             case "ruby.ignore-lines-before-ruby-shebang":
                 return IGNORE_LINES_BEFORE_RUBY_SHEBANG;
             case "ruby.syntax-check":